simplified part validator
