--- modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/snapshot/AbstractSnapshotPartitionsVerifyHandler.java
@@ -1,267 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.ignite.internal.processors.cache.persistence.snapshot;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.DirectoryStream;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
-import org.apache.ignite.IgniteCheckedException;
-import org.apache.ignite.IgniteException;
-import org.apache.ignite.IgniteLogger;
-import org.apache.ignite.internal.GridComponent;
-import org.apache.ignite.internal.GridKernalContext;
-import org.apache.ignite.internal.managers.encryption.EncryptionCacheKeyProvider;
-import org.apache.ignite.internal.managers.encryption.GroupKey;
-import org.apache.ignite.internal.managers.encryption.GroupKeyEncrypted;
-import org.apache.ignite.internal.processors.cache.GridCacheSharedContext;
-import org.apache.ignite.internal.processors.cache.StoredCacheData;
-import org.apache.ignite.internal.processors.cache.persistence.file.FilePageStore;
-import org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager;
-import org.apache.ignite.internal.processors.cache.verify.PartitionKeyV2;
-import org.apache.ignite.internal.util.typedef.F;
-import org.apache.ignite.internal.util.typedef.internal.CU;
-import org.apache.ignite.internal.util.typedef.internal.U;
-import org.jetbrains.annotations.Nullable;
-
-import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.CACHE_DATA_FILENAME;
-import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.cacheDirectories;
-import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.cacheGroupName;
-import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.cachePartitionFiles;
-import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.partId;
-import static org.apache.ignite.internal.processors.cache.persistence.partstate.GroupPartitionId.getTypeByPartId;
-import static org.apache.ignite.internal.processors.cache.persistence.snapshot.IgniteSnapshotManager.databaseRelativePath;
-
-/**
- * Snapshot handler base for checking snapshot partitions consistency.
- *
- * @param <R> Partition check result.
- */
-abstract class AbstractSnapshotPartitionsVerifyHandler<R> implements SnapshotHandler<Map<PartitionKeyV2, R>> {
-    /** Shared context. */
-    protected final GridCacheSharedContext<?, ?> cctx;
-
-    /** Logger. */
-    protected final IgniteLogger log;
-
-    /**
-     * @param cctx Shared context.
-     */
-    AbstractSnapshotPartitionsVerifyHandler(GridCacheSharedContext<?, ?> cctx) {
-        this.cctx = cctx;
-
-        log = cctx.logger(getClass());
-    }
-
-    /** {@inheritDoc} */
-    @Override public Map<PartitionKeyV2, R> invoke(SnapshotHandlerContext hndCtx) throws IgniteCheckedException {
-        if (!hndCtx.snapshotDirectory().exists())
-            throw new IgniteCheckedException("Snapshot directory doesn't exists: " + hndCtx.snapshotDirectory());
-
-        SnapshotMetadata meta = hndCtx.metadata();
-
-        Set<Integer> grps = F.isEmpty(hndCtx.groups()) ? new HashSet<>(meta.partitions().keySet()) :
-            hndCtx.groups().stream().map(CU::cacheId).collect(Collectors.toSet());
-
-        Set<File> partFiles = new HashSet<>();
-
-        Map<Integer, File> grpDirs = new HashMap<>();
-
-        for (File dir : cacheDirectories(new File(hndCtx.snapshotDirectory(), databaseRelativePath(meta.folderName())), name -> true)) {
-            int grpId = CU.cacheId(cacheGroupName(dir));
-
-            if (!grps.remove(grpId))
-                continue;
-
-            Set<Integer> parts = meta.partitions().get(grpId) == null ? Collections.emptySet() :
-                new HashSet<>(meta.partitions().get(grpId));
-
-            for (File part : cachePartitionFiles(dir)) {
-                int partId = partId(part.getName());
-
-                if (!parts.remove(partId))
-                    continue;
-
-                partFiles.add(part);
-            }
-
-            if (!parts.isEmpty()) {
-                throw new IgniteException("Snapshot data doesn't contain required cache group partition " +
-                    "[grpId=" + grpId + ", snpName=" + meta.snapshotName() + ", consId=" + meta.consistentId() +
-                    ", missed=" + parts + ", meta=" + meta + ']');
-            }
-
-            grpDirs.put(grpId, dir);
-        }
-
-        if (!grps.isEmpty()) {
-            throw new IgniteException("Snapshot data doesn't contain required cache groups " +
-                "[grps=" + grps + ", snpName=" + meta.snapshotName() + ", consId=" + meta.consistentId() +
-                ", meta=" + meta + ']');
-        }
-
-        Map<PartitionKeyV2, R> res = new ConcurrentHashMap<>();
-
-        IgniteSnapshotManager snpMgr = cctx.snapshotMgr();
-
-        GridKernalContext opCtx = type() == SnapshotHandlerType.CREATE ? cctx.kernalContext() :
-            cctx.snapshotMgr().createStandaloneKernalContext(hndCtx.snapshotDirectory(), hndCtx.metadata().folderName());
-
-        FilePageStoreManager storeMgr = (FilePageStoreManager)cctx.pageStore();
-
-        EncryptionCacheKeyProvider snpEncrKeyProvider = new SnapshotEncryptionKeyProvider(cctx.kernalContext(), grpDirs);
-
-        if (type() == SnapshotHandlerType.RESTORE) {
-            for (GridComponent comp : opCtx)
-                comp.start();
-        }
-
-        try {
-            U.doInParallel(
-                snpMgr.snapshotExecutorService(),
-                partFiles,
-                part -> {
-                    String grpName = cacheGroupName(part.getParentFile());
-                    int grpId = CU.cacheId(grpName);
-                    int partId = partId(part.getName());
-
-                    PartitionKeyV2 partKey = new PartitionKeyV2(grpId, partId, grpName);
-
-                    try (FilePageStore pageStore =
-                             (FilePageStore)storeMgr.getPageStoreFactory(grpId, snpEncrKeyProvider.getActiveKey(grpId) != null ?
-                                 snpEncrKeyProvider : null).createPageStore(getTypeByPartId(partId), part::toPath, val -> {})
-                    ) {
-                        R partRes = validatePartition(hndCtx, opCtx, partKey, pageStore);
-
-                        if (partRes != null)
-                            res.put(partKey, partRes);
-                    }
-                    catch (IOException e) {
-                        throw new IgniteCheckedException(e);
-                    }
-
-                    return null;
-                }
-            );
-        }
-        catch (Throwable t) {
-            log.error("Error executing handler: ", t);
-
-            throw t;
-        }
-        finally {
-            if (type() == SnapshotHandlerType.RESTORE) {
-                for (GridComponent comp : opCtx)
-                    comp.stop(true);
-            }
-        }
-
-        return res;
-    }
-
-    /**
-     * Validates certain partition. Including the meta and index.
-     *
-     * @param hndCtx Snapshot handler context.
-     * @param opCtx Snapshot operation context.
-     * @param partKey Partition key.
-     * @param pageStore Store to read and check from.
-     * @return A result for the partition or {@code null}.
-     */
-    protected abstract R validatePartition(
-        SnapshotHandlerContext hndCtx,
-        GridKernalContext opCtx,
-        PartitionKeyV2 partKey,
-        FilePageStore pageStore
-    ) throws IgniteCheckedException;
-
-    /**
-     * Provides encryption keys stored within snapshot.
-     * <p>
-     * To restore an encrypted snapshot, we have to read the keys it was encrypted with. The better place for the is
-     * Metastore. But it is currently unreadable as simple structure. Once it is done, we should move snapshot
-     * encryption keys there.
-     */
-    private static class SnapshotEncryptionKeyProvider implements EncryptionCacheKeyProvider {
-        /** Kernal context */
-        private final GridKernalContext ctx;
-
-        /** Data dirs of snapshot's caches by group id. */
-        private final Map<Integer, File> grpDirs;
-
-        /** Encryption keys loaded from snapshot. */
-        private final ConcurrentHashMap<Integer, GroupKey> decryptedKeys = new ConcurrentHashMap<>();
-
-        /**
-         * Constructor.
-         *
-         * @param ctx     Kernal context.
-         * @param grpDirs Data dirictories of cache groups by id.
-         */
-        private SnapshotEncryptionKeyProvider(GridKernalContext ctx, Map<Integer, File> grpDirs) {
-            this.ctx = ctx;
-            this.grpDirs = grpDirs;
-        }
-
-        /** {@inheritDoc} */
-        @Override public @Nullable GroupKey getActiveKey(int grpId) {
-            return decryptedKeys.computeIfAbsent(grpId, id -> {
-                GroupKey grpKey = null;
-
-                try (DirectoryStream<Path> ds = Files.newDirectoryStream(grpDirs.get(grpId).toPath(),
-                    p -> Files.isRegularFile(p) && p.toString().endsWith(CACHE_DATA_FILENAME))) {
-                    for (Path p : ds) {
-                        StoredCacheData cacheData = ctx.cache().configManager().readCacheData(p.toFile());
-
-                        GroupKeyEncrypted grpKeyEncrypted = cacheData.groupKeyEncrypted();
-
-                        if (grpKeyEncrypted == null)
-                            return null;
-
-                        if (grpKey == null)
-                            grpKey = new GroupKey(grpKeyEncrypted.id(), ctx.config().getEncryptionSpi().decryptKey(grpKeyEncrypted.key()));
-                        else {
-                            assert grpKey.equals(new GroupKey(grpKeyEncrypted.id(),
-                                ctx.config().getEncryptionSpi().decryptKey(grpKeyEncrypted.key())));
-                        }
-                    }
-
-                    return grpKey;
-                }
-                catch (Exception e) {
-                    throw new IgniteException("Unable to extract ciphered encryption key of cache group " + id + '.', e);
-                }
-            });
-        }
-
-        /** {@inheritDoc} */
-        @Override public @Nullable GroupKey groupKey(int grpId, int keyId) {
-            GroupKey key = getActiveKey(grpId);
-
-            return key != null && key.id() == keyId ? key : null;
-        }
-    }
-}
