+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/snapshot/SnapshotPartitionsVerifyHandler.java
@@ -17,19 +17,38 @@
 
 package org.apache.ignite.internal.processors.cache.persistence.snapshot;
 
+import java.io.File;
+import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
+import java.nio.file.DirectoryStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
 import org.apache.ignite.IgniteCheckedException;
+import org.apache.ignite.IgniteException;
+import org.apache.ignite.IgniteLogger;
 import org.apache.ignite.cluster.ClusterNode;
+import org.apache.ignite.internal.GridComponent;
 import org.apache.ignite.internal.GridKernalContext;
+import org.apache.ignite.internal.managers.encryption.EncryptionCacheKeyProvider;
+import org.apache.ignite.internal.managers.encryption.GroupKey;
+import org.apache.ignite.internal.managers.encryption.GroupKeyEncrypted;
 import org.apache.ignite.internal.processors.cache.GridCacheSharedContext;
+import org.apache.ignite.internal.processors.cache.StoredCacheData;
 import org.apache.ignite.internal.processors.cache.distributed.dht.topology.GridDhtPartitionState;
+import org.apache.ignite.internal.processors.cache.persistence.CacheDataRow;
 import org.apache.ignite.internal.processors.cache.persistence.file.FilePageStore;
+import org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager;
 import org.apache.ignite.internal.processors.cache.persistence.metastorage.MetaStorage;
 import org.apache.ignite.internal.processors.cache.persistence.tree.io.PageIO;
 import org.apache.ignite.internal.processors.cache.persistence.tree.io.PagePartitionMetaIO;
@@ -38,22 +57,42 @@ import org.apache.ignite.internal.processors.cache.verify.PartitionHashRecordV2;
 import org.apache.ignite.internal.processors.cache.verify.PartitionKeyV2;
 import org.apache.ignite.internal.util.GridStringBuilder;
 import org.apache.ignite.internal.util.GridUnsafe;
+import org.apache.ignite.internal.util.lang.GridIterator;
+import org.apache.ignite.internal.util.typedef.F;
+import org.apache.ignite.internal.util.typedef.internal.CU;
+import org.apache.ignite.internal.util.typedef.internal.U;
+import org.jetbrains.annotations.Nullable;
 
 import static org.apache.ignite.internal.pagemem.PageIdAllocator.FLAG_DATA;
 import static org.apache.ignite.internal.pagemem.PageIdAllocator.FLAG_IDX;
 import static org.apache.ignite.internal.pagemem.PageIdAllocator.INDEX_PARTITION;
 import static org.apache.ignite.internal.processors.cache.distributed.dht.topology.GridDhtPartitionState.OWNING;
 import static org.apache.ignite.internal.processors.cache.distributed.dht.topology.GridDhtPartitionState.fromOrdinal;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.CACHE_DATA_FILENAME;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.cacheDirectories;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.cacheGroupName;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.cachePartitionFiles;
+import static org.apache.ignite.internal.processors.cache.persistence.file.FilePageStoreManager.partId;
+import static org.apache.ignite.internal.processors.cache.persistence.partstate.GroupPartitionId.getTypeByPartId;
+import static org.apache.ignite.internal.processors.cache.persistence.snapshot.IgniteSnapshotManager.databaseRelativePath;
 import static org.apache.ignite.internal.processors.cache.verify.IdleVerifyUtility.calculatePartitionHash;
 import static org.apache.ignite.internal.processors.cache.verify.IdleVerifyUtility.checkPartitionsPageCrcSum;
 
 /**
  * Default snapshot restore handler for checking snapshot partitions consistency.
  */
-public class SnapshotPartitionsVerifyHandler extends AbstractSnapshotPartitionsVerifyHandler<PartitionHashRecordV2> {
-    /**  @param cctx Shared context. */
-    SnapshotPartitionsVerifyHandler(GridCacheSharedContext<?, ?> cctx) {
-        super(cctx);
+public class SnapshotPartitionsVerifyHandler implements SnapshotHandler<Map<PartitionKeyV2, PartitionHashRecordV2>> {
+    /** Shared context. */
+    protected final GridCacheSharedContext<?, ?> cctx;
+
+    /** Logger. */
+    private final IgniteLogger log;
+
+    /** @param cctx Shared context. */
+    public SnapshotPartitionsVerifyHandler(GridCacheSharedContext<?, ?> cctx) {
+        this.cctx = cctx;
+
+        log = cctx.logger(getClass());
     }
 
     /** {@inheritDoc} */
@@ -62,61 +101,170 @@ public class SnapshotPartitionsVerifyHandler extends AbstractSnapshotPartitionsV
     }
 
     /** {@inheritDoc} */
-    @Override protected PartitionHashRecordV2 validatePartition(
-        SnapshotHandlerContext hndCtx,
-        GridKernalContext opCtx,
-        PartitionKeyV2 partKey,
-        FilePageStore pageStore
+    @Override public Map<PartitionKeyV2, PartitionHashRecordV2> invoke(SnapshotHandlerContext opCtx) throws IgniteCheckedException {
+        if (!opCtx.snapshotDirectory().exists())
+            throw new IgniteCheckedException("Snapshot directory doesn't exists: " + opCtx.snapshotDirectory());;
+
+        SnapshotMetadata meta = opCtx.metadata();
+
+        Set<Integer> grps = F.isEmpty(opCtx.groups()) ? new HashSet<>(meta.partitions().keySet()) :
+            opCtx.groups().stream().map(CU::cacheId).collect(Collectors.toSet());
+
+        Set<File> partFiles = new HashSet<>();
+
+        Map<Integer, File> grpDirs = new HashMap<>();
+
+        for (File dir : cacheDirectories(new File(opCtx.snapshotDirectory(), databaseRelativePath(meta.folderName())), name -> true)) {
+            int grpId = CU.cacheId(cacheGroupName(dir));
+
+            if (!grps.remove(grpId))
+                continue;
+
+            Set<Integer> parts = meta.partitions().get(grpId) == null ? Collections.emptySet() :
+                new HashSet<>(meta.partitions().get(grpId));
+
+            for (File part : cachePartitionFiles(dir)) {
+                int partId = partId(part.getName());
+
+                if (!parts.remove(partId))
+                    continue;
+
+                partFiles.add(part);
+            }
+
+            if (!parts.isEmpty()) {
+                throw new IgniteException("Snapshot data doesn't contain required cache group partition " +
+                    "[grpId=" + grpId + ", snpName=" + meta.snapshotName() + ", consId=" + meta.consistentId() +
+                    ", missed=" + parts + ", meta=" + meta + ']');
+            }
+
+            grpDirs.put(grpId, dir);
+        }
+
+        if (!grps.isEmpty()) {
+            throw new IgniteException("Snapshot data doesn't contain required cache groups " +
+                "[grps=" + grps + ", snpName=" + meta.snapshotName() + ", consId=" + meta.consistentId() +
+                ", meta=" + meta + ']');
+        }
+
+        Map<PartitionKeyV2, PartitionHashRecordV2> res = new ConcurrentHashMap<>();
+        ThreadLocal<ByteBuffer> buff = ThreadLocal.withInitial(() -> ByteBuffer.allocateDirect(meta.pageSize())
+            .order(ByteOrder.nativeOrder()));
+
+        IgniteSnapshotManager snpMgr = cctx.snapshotMgr();
+
+        GridKernalContext snpCtx = snpMgr.createStandaloneKernalContext(opCtx.snapshotDirectory(), meta.folderName());
+
+        FilePageStoreManager storeMgr = (FilePageStoreManager)cctx.pageStore();
+
+        EncryptionCacheKeyProvider snpEncrKeyProvider = new SnapshotEncryptionKeyProvider(cctx.kernalContext(), grpDirs);
+
+        for (GridComponent comp : snpCtx)
+            comp.start();
+
+        try {
+            U.doInParallel(
+                snpMgr.snapshotExecutorService(),
+                partFiles,
+                part -> {
+                    String grpName = cacheGroupName(part.getParentFile());
+                    int grpId = CU.cacheId(grpName);
+                    int partId = partId(part.getName());
+
+                    try (FilePageStore pageStore =
+                             (FilePageStore)storeMgr.getPageStoreFactory(grpId, snpEncrKeyProvider.getActiveKey(grpId) != null ?
+                                 snpEncrKeyProvider : null).createPageStore(getTypeByPartId(partId), part::toPath, val -> {})
+                    ) {
+                        if (partId == INDEX_PARTITION) {
+                            checkPartitionsPageCrcSum(() -> pageStore, INDEX_PARTITION, FLAG_IDX);
+
+                            return null;
+                        }
+
+                        if (grpId == MetaStorage.METASTORAGE_CACHE_ID) {
+                            checkPartitionsPageCrcSum(() -> pageStore, partId, FLAG_DATA);
+
+                            return null;
+                        }
+
+                        ByteBuffer pageBuff = buff.get();
+                        pageBuff.clear();
+                        pageStore.read(0, pageBuff, true);
+
+                        long pageAddr = GridUnsafe.bufferAddress(pageBuff);
+
+                        PagePartitionMetaIO io = PageIO.getPageIO(pageBuff);
+                        GridDhtPartitionState partState = fromOrdinal(io.getPartitionState(pageAddr));
+
+                        if (partState != OWNING) {
+                            throw new IgniteCheckedException("Snapshot partitions must be in the OWNING " +
+                                "state only: " + partState);
+                        }
+
+                        long updateCntr = io.getUpdateCounter(pageAddr);
+                        long size = io.getSize(pageAddr);
+
+                        if (log.isDebugEnabled()) {
+                            log.debug("Partition [grpId=" + grpId
+                                + ", id=" + partId
+                                + ", counter=" + updateCntr
+                                + ", size=" + size + "]");
+                        }
+
+                        // Snapshot partitions must always be in OWNING state.
+                        // There is no `primary` partitions for snapshot.
+                        PartitionKeyV2 key = new PartitionKeyV2(grpId, partId, grpName);
+
+                        PartitionHashRecordV2 hash = partHash(key,
+                            updateCntr,
+                            meta.consistentId(),
+                            GridDhtPartitionState.OWNING,
+                            false,
+                            size,
+                            snpMgr.partitionRowIterator(snpCtx, grpName, partId, pageStore));
+
+                        res.put(key, hash);
+                    }
+                    catch (IOException e) {
+                        throw new IgniteCheckedException(e);
+                    }
+
+                    return null;
+                }
+            );
+        }
+        catch (Throwable t) {
+            log.error("Error executing handler: ", t);
+
+            throw t;
+        }
+        finally {
+            for (GridComponent comp : snpCtx)
+                comp.stop(true);
+        }
+
+        return res;
+    }
+
+    /** */
+    protected PartitionHashRecordV2 partHash(
+        PartitionKeyV2 key,
+        Object updCntr,
+        Object consId,
+        GridDhtPartitionState state,
+        boolean isPrimary,
+        long partSize,
+        GridIterator<CacheDataRow> it
     ) throws IgniteCheckedException {
-        if (partKey.partitionId() == INDEX_PARTITION) {
-            checkPartitionsPageCrcSum(() -> pageStore, INDEX_PARTITION, FLAG_IDX);
+        PartitionHashRecordV2 hash = calculatePartitionHash(key,
+            updCntr,
+            consId,
+            state,
+            isPrimary,
+            partSize,
+            it);
 
-            return null;
-        }
-
-        if (partKey.groupId() == MetaStorage.METASTORAGE_CACHE_ID) {
-            checkPartitionsPageCrcSum(() -> pageStore, partKey.partitionId(), FLAG_DATA);
-
-            return null;
-        }
-
-        ThreadLocal<ByteBuffer> buff =
-            ThreadLocal.withInitial(() -> ByteBuffer.allocateDirect(hndCtx.metadata().pageSize()).order(ByteOrder.nativeOrder()));
-        ByteBuffer pageBuff = buff.get();
-        pageBuff.clear();
-        pageStore.read(0, pageBuff, true);
-
-        long pageAddr = GridUnsafe.bufferAddress(pageBuff);
-
-        PagePartitionMetaIO io = PageIO.getPageIO(pageBuff);
-        GridDhtPartitionState partState = fromOrdinal(io.getPartitionState(pageAddr));
-
-        if (partState != OWNING) {
-            throw new IgniteCheckedException("Snapshot partitions must be in the OWNING " +
-                "state only: " + partState);
-        }
-
-        long updateCntr = io.getUpdateCounter(pageAddr);
-        long size = io.getSize(pageAddr);
-
-        if (log.isDebugEnabled()) {
-            log.debug("Partition [grpId=" + partKey.groupId()
-                + ", id=" + partKey.partitionId()
-                + ", counter=" + updateCntr
-                + ", size=" + size + "]");
-        }
-
-        // Snapshot partitions must always be in OWNING state.
-        // There is no `primary` partitions for snapshot.
-        PartitionHashRecordV2 hash = calculatePartitionHash(partKey,
-            updateCntr,
-            hndCtx.metadata().consistentId(),
-            GridDhtPartitionState.OWNING,
-            false,
-            size,
-            cctx.snapshotMgr().partitionRowIterator(opCtx, partKey.groupName(), partKey.partitionId(), pageStore));
-
-        assert hash != null : "OWNING must have hash: " + partKey;
+        assert hash != null : "OWNING must have hash: " + key;
 
         return hash;
     }
@@ -148,5 +296,72 @@ public class SnapshotPartitionsVerifyHandler extends AbstractSnapshotPartitionsV
         verifyResult.print(buf::a, true);
 
         throw new IgniteCheckedException(buf.toString());
+    }
+
+    /**
+     * Provides encryption keys stored within snapshot.
+     * <p>
+     * To restore an encrypted snapshot, we have to read the keys it was encrypted with. The better place for the is
+     * Metastore. But it is currently unreadable as simple structure. Once it is done, we should move snapshot
+     * encryption keys there.
+     */
+    private static class SnapshotEncryptionKeyProvider implements EncryptionCacheKeyProvider {
+        /** Kernal context */
+        private final GridKernalContext ctx;
+
+        /** Data dirs of snapshot's caches by group id. */
+        private final Map<Integer, File> grpDirs;
+
+        /** Encryption keys loaded from snapshot. */
+        private final ConcurrentHashMap<Integer, GroupKey> decryptedKeys = new ConcurrentHashMap<>();
+
+        /**
+         * Constructor.
+         *
+         * @param ctx     Kernal context.
+         * @param grpDirs Data dirictories of cache groups by id.
+         */
+        private SnapshotEncryptionKeyProvider(GridKernalContext ctx, Map<Integer, File> grpDirs) {
+            this.ctx = ctx;
+            this.grpDirs = grpDirs;
+        }
+
+        /** {@inheritDoc} */
+        @Override public @Nullable GroupKey getActiveKey(int grpId) {
+            return decryptedKeys.computeIfAbsent(grpId, id -> {
+                GroupKey grpKey = null;
+
+                try (DirectoryStream<Path> ds = Files.newDirectoryStream(grpDirs.get(grpId).toPath(),
+                    p -> Files.isRegularFile(p) && p.toString().endsWith(CACHE_DATA_FILENAME))) {
+                    for (Path p : ds) {
+                        StoredCacheData cacheData = ctx.cache().configManager().readCacheData(p.toFile());
+
+                        GroupKeyEncrypted grpKeyEncrypted = cacheData.groupKeyEncrypted();
+
+                        if (grpKeyEncrypted == null)
+                            return null;
+
+                        if (grpKey == null)
+                            grpKey = new GroupKey(grpKeyEncrypted.id(), ctx.config().getEncryptionSpi().decryptKey(grpKeyEncrypted.key()));
+                        else {
+                            assert grpKey.equals(new GroupKey(grpKeyEncrypted.id(),
+                                ctx.config().getEncryptionSpi().decryptKey(grpKeyEncrypted.key())));
+                        }
+                    }
+
+                    return grpKey;
+                }
+                catch (Exception e) {
+                    throw new IgniteException("Unable to extract ciphered encryption key of cache group " + id + '.', e);
+                }
+            });
+        }
+
+        /** {@inheritDoc} */
+        @Override public @Nullable GroupKey groupKey(int grpId, int keyId) {
+            GroupKey key = getActiveKey(grpId);
+
+            return key != null && key.id() == keyId ? key : null;
+        }
     }
 }
